Update outline.md
line 30 "analogy is a roadtrip" - sticking with 'cookbook,' I think of this as fodder for planning a series of dinner parties. The cookbook is intended as a guide, and for iterative use, not as a letter-of-the-law formula that is sequential and intolerant of deviation. I think of this as open to interpretation and customization.