Chopping community development content
We’ve made the decision to confine the PIE Cookbook to Accelerator
Management. We are working under the assumption that the community
already exists and we will lightly touch on that a number of times.
Given that building out these Community Management guidelines would
take significant time and effort — and be a book in and of themselves —
we are striking them from the document.